test(batch-costing): split page scenarios
